#15: Refactored report field formatting to use annotations and not so hack-tastically reflection-happy
